Refactored core module.
