+++ core/src/core/mod.rs
@@ -1,0 +1,5 @@
+pub mod screen;
+pub mod stack;
+pub mod workspace;
+pub mod workspaces;
+pub mod rational_rect;
