+++ core/src/core/rational_rect.rs
@@ -1,0 +1,4 @@
+
+
+#[derive(Clone, Copy, Debug)]
+pub struct RationalRect(pub f32, pub f32, pub f32, pub f32);
