+++ core/src/core/workspaces.rs
@@ -5,313 +5,10 @@ use config::GeneralConfig;
 use window_manager::ScreenDetail;
 use window_system::{ Window, WindowSystem };
 use layout::{Layout, LayoutMessage};
-
-#[derive(Clone, Copy, Debug)]
-pub struct RationalRect(pub f32, pub f32, pub f32, pub f32);
-
-/// Handles focus tracking on a workspace.
-/// `focus` keeps track of the focused window's id
-/// and `up` and `down` are the windows above or
-/// below the focus stack respectively.
-#[derive(Clone, PartialEq, Eq)]
-pub struct Stack<T> {
-    pub focus: T,
-    pub up:    Vec<T>,
-    pub down:  Vec<T>
-}
-
-impl<T: Clone + Eq> Stack<T> {
-    /// Create a new stack with the given values
-    pub fn new(f: T, up: Vec<T>, down: Vec<T>) -> Stack<T> {
-        Stack {
-            focus: f,
-            up:    up,
-            down:  down
-        }
-    }
-
-    /// Create a new stack with only the given element
-    /// as the focused one and initialize the rest to empty.
-    pub fn from_element(t: T) -> Stack<T> {
-        Stack {
-            focus: t,
-            up:    Vec::new(),
-            down:  Vec::new()
-        }
-    }
-
-    /// Add a new element to the stack
-    /// and automatically focus it.
-    pub fn add(&self, t: T) -> Stack<T> {
-        Stack {
-            focus: t,
-            up: self.up.clone(),
-            down: self.down.clone() + (vec!(self.focus.clone()).as_slice())
-        }
-    }
-
-    /// Flatten the stack into a list
-    pub fn integrate(&self) -> Vec<T> {
-        self.up.iter()
-            .rev()
-            .chain((vec!(self.focus.clone())).iter())
-            .chain(self.down.iter())
-            .map(|x| x.clone())
-            .collect()
-    }
-
-    /// Filter the stack to retain only windows
-    /// that yield true in the given filter function
-    pub fn filter<F>(&self, f: F) -> Option<Stack<T>> where F : Fn(&T) -> bool {
-        let lrs : Vec<T> = (vec!(self.focus.clone()) + self.down.as_slice()).iter()
-            .filter(|&x| f(x))
-            .map(|x| x.clone())
-            .collect();
-
-        if lrs.len() > 0 {
-            let first : T         = lrs[0].clone();
-            let rest : Vec<T>     = lrs.iter().skip(1).map(|x| x.clone()).collect();
-            let filtered : Vec<T> = self.up.iter()
-                .filter(|&x| f(x))
-                .map(|x| x.clone())
-                .collect();
-            let stack : Stack<T>  = Stack::<T>::new(first, filtered, rest);
-
-            Some(stack)
-        } else {
-            let filtered : Vec<T> = self.up.iter().map(|x| x.clone()).filter(|x| f(x)).collect();
-            if filtered.len() > 0 {
-                let first : T        = filtered[0].clone();
-                let rest : Vec<T>    = filtered.iter().skip(1).map(|x| x.clone()).collect();
-
-                Some(Stack::<T>::new(first, rest, Vec::new()))
-            } else {
-                None
-            }
-        }
-    }
-
-    /// Move the focus to the next element in the `up` list
-    pub fn focus_up(&self) -> Stack<T> {
-        if self.up.is_empty() {
-            let tmp : Vec<T> = (vec!(self.focus.clone()) + self.down.as_slice()).into_iter()
-                .rev()
-                .collect();
-            let xs : Vec<T> = tmp.iter()
-                .skip(1)
-                .map(|x| x.clone())
-                .collect();
-
-            Stack::<T>::new(tmp[0].clone(), xs, Vec::new())
-        } else {
-            let down = (vec!(self.focus.clone())) + self.down.as_slice();
-            let up   = self.up.iter().skip(1).map(|x| x.clone()).collect();
-            Stack::<T>::new(self.up[0].clone(), up, down)
-        }
-    }
-
-    /// Move the focus down
-    pub fn focus_down(&self) -> Stack<T> {
-        self.reverse().focus_up().reverse()
-    }
-
-    pub fn swap_up(&self) -> Stack<T> {
-        if self.up.is_empty() {
-            Stack::<T>::new(self.focus.clone(), self.down.iter().rev().map(|x| x.clone()).collect(), Vec::new())
-        } else {
-            let x = self.up[0].clone();
-            let xs = self.up.iter().skip(1).map(|x| x.clone()).collect();
-            let rs = vec!(x) + self.down.as_slice();
-            Stack::<T>::new(self.focus.clone(), xs, rs)
-        }
-    }
-
-    pub fn swap_down(&self) -> Stack<T> {
-        self.reverse().swap_up().reverse()
-    }
-
-    pub fn swap_master(&self) -> Stack<T> {
-        if self.up.is_empty() {
-            return self.clone();
-        }
-
-        let r : Vec<T>  = self.up.iter()
-            .rev()
-            .map(|x| x.clone())
-            .collect();
-        let x : T       = r[0].clone();
-        let xs : Vec<T> = r.iter()
-            .skip(1)
-            .map(|x| x.clone())
-            .collect();
-        let rs : Vec<T> = xs + (vec!(x) + self.down.as_slice()).as_slice();
-
-        Stack::<T>::new(self.focus.clone(), Vec::new(), rs)
-    }
-
-    /// Reverse the stack by exchanging
-    /// the `up` and `down` lists
-    pub fn reverse(&self) -> Stack<T> {
-        Stack::<T>::new(self.focus.clone(), self.down.clone(), self.up.clone())
-    }
-
-    /// Return the number of elements tracked by the stack
-    pub fn len(&self) -> usize {
-        1 + self.up.len() + self.down.len()
-    }
-
-    /// Checks if the given window is tracked by the stack
-    pub fn contains(&self, window: T) -> bool {
-        self.focus == window || self.up.contains(&window) || self.down.contains(&window)
-    }
-}
-
-/// Represents a single workspace with a `tag` (name),
-/// `id`, a `layout` and a `stack` for all windows
-pub struct Workspace {
-    pub id:     u32,
-    pub tag:    String,
-    pub layout: Box<Layout>,
-    pub stack:  Option<Stack<Window>>
-}
-
-impl Clone for Workspace {
-    fn clone(&self) -> Workspace {
-        Workspace {
-            id: self.id,
-            tag: self.tag.clone(),
-            layout: self.layout.copy(),
-            stack: self.stack.clone()
-        }
-    }
-}
-
-impl Workspace {
-    /// Create a new workspace
-    pub fn new(id: u32, tag: String, layout: Box<Layout>, stack: Option<Stack<Window>>) -> Workspace {
-        Workspace {
-            id: id,
-            tag: tag,
-            layout: layout,
-            stack: stack
-        }
-    }
-
-    /// Add a new window to the workspace by adding it to the stack.
-    /// If the stack doesn't exist yet, create one.
-    pub fn add(&self, window: Window) -> Workspace {
-        Workspace::new(
-            self.id,
-            self.tag.clone(),
-            self.layout.copy(),
-            Some(self.stack.clone().map_or(Stack::from_element(window), |s| s.add(window))))
-    }
-
-    /// Returns the number of windows contained in this workspace
-    pub fn len(&self) -> usize {
-        self.stack.clone().map_or(0, |x| x.len())
-    }
-
-    /// Checks if the workspace contains the given window
-    pub fn contains(&self, window: Window) -> bool {
-        self.stack.clone().map_or(false, |x| x.contains(window))
-    }
-
-    pub fn windows(&self) -> Vec<Window> {
-        self.stack.clone().map_or(Vec::new(), |s| s.integrate())
-    }
-
-    pub fn peek(&self) -> Option<Window> {
-        self.stack.clone().map(|s| s.focus)
-    }
-
-    pub fn map<F>(&self, f: F) -> Workspace where F : Fn(Stack<Window>) -> Stack<Window> {
-        Workspace::new(self.id, self.tag.clone(), self.layout.copy(),
-        self.stack.clone().map(|x| f(x)))
-    }
-
-    pub fn map_option<F>(&self, f: F) -> Workspace where F : Fn(Stack<Window>) -> Option<Stack<Window>> {
-        Workspace::new(self.id, self.tag.clone(), self.layout.copy(),
-        self.stack.clone().map_or(None, |x| f(x)))
-    }
-
-    pub fn map_or<F>(&self, default: Stack<Window>, f: F) -> Workspace where F : Fn(Stack<Window>) -> Stack<Window> {
-        Workspace::new(self.id, self.tag.clone(), self.layout.copy(),
-        Some(self.stack.clone().map_or(default, |x| f(x))))
-    }
-
-    pub fn send_layout_message(&self, message: LayoutMessage, window_system: &WindowSystem,
-                                   config: &GeneralConfig) -> Workspace {
-        let mut layout = self.layout.copy();
-        layout.apply_message(message, window_system, &self.stack, config);
-        Workspace::new(self.id, self.tag.clone(), layout, self.stack.clone())
-    }
-}
-
-pub struct Screen {
-    pub workspace:     Workspace,
-    pub screen_id:     u32,
-    pub screen_detail: ScreenDetail
-}
-
-impl Clone for Screen {
-    fn clone(&self) -> Screen {
-        Screen {
-            workspace: self.workspace.clone(),
-            screen_id: self.screen_id,
-            screen_detail: self.screen_detail.clone()
-        }
-    }
-}
-
-impl Screen {
-    /// Create a new screen for the given workspace
-    /// and the given dimensions
-    pub fn new(workspace: Workspace, screen_id: u32, screen_detail: ScreenDetail) -> Screen {
-        Screen {
-            workspace: workspace,
-            screen_id: screen_id,
-            screen_detail: screen_detail
-        }
-    }
-
-    /// Checks if the screen's workspace contains
-    /// the given window
-    pub fn contains(&self, window: Window) -> bool {
-        self.workspace.contains(window)
-    }
-
-    /// Returns the number of windows in the
-    /// screen's workspace
-    pub fn len(&self) -> usize {
-        self.workspace.len()
-    }
-
-    pub fn windows(&self) -> Vec<Window> {
-        self.workspace.windows()
-    }
-
-    pub fn map_workspace<F>(&self, f: F) -> Screen where F : Fn(Workspace) -> Workspace {
-        Screen::new(f(self.workspace.clone()), self.screen_id, self.screen_detail)
-    }
-
-    pub fn map<F>(&self, f: F) -> Screen where F : Fn(Stack<Window>) -> Stack<Window> {
-        Screen::new(self.workspace.map(f), self.screen_id, self.screen_detail)
-    }
-
-    pub fn map_option<F>(&self, f: F) -> Screen where F : Fn(Stack<Window>) -> Option<Stack<Window>> {
-        Screen::new(self.workspace.map_option(f), self.screen_id, self.screen_detail)
-    }
-
-    pub fn map_or<F>(&self, default: Stack<Window>, f: F) -> Screen where F : Fn(Stack<Window>) -> Stack<Window> {
-        Screen::new(self.workspace.map_or(default, f), self.screen_id, self.screen_detail)
-    }
-
-    pub fn send_layout_message(&self, message: LayoutMessage, window_system: &WindowSystem,
-                                   config: &GeneralConfig) -> Screen {
-        Screen::new(self.workspace.send_layout_message(message, window_system, config), self.screen_id, self.screen_detail)
-    }
-}
+use core::screen::Screen;
+use core::rational_rect::RationalRect;
+use core::workspace::Workspace;
+use core::stack::Stack;
 
 pub struct Workspaces {
     /// The currently focused and visible screen
@@ -622,7 +319,7 @@ impl Workspaces {
 
     /// Retrieve the currently focused workspace's id
     pub fn current_tag(&self) -> u32 {
-        self.current.workspace.id as u32
+        self.current.workspace.id
     }
 
     /// Retrieve the tag of the workspace the given window
@@ -632,7 +329,7 @@ impl Workspaces {
         debug!("trying to find tag of workspace with window {}", window);
         self.workspaces().iter()
             .filter(|x| x.contains(window))
-            .map(|x| x.id as u32)
+            .map(|x| x.id)
             .nth(0)
     }
