+++ core/tests/core.rs
@@ -1,9 +1,10 @@
 #![feature(collections)]
 
-use wtftw_core::core::*;
-use wtftw_core::layout::FullLayout;
-
 extern crate wtftw_core;
+
+use wtftw_core::layout::FullLayout;
+use wtftw_core::core::stack::Stack;
+use wtftw_core::core::workspace::Workspace;
 
 #[test]
 fn stack_add() {
